agent/requirements.txt
openai
pydantic>=2.0
click
python-dotenv